MulMeokNyang/AI/requirements.txt
colormath==3.0.0
Flask==3.0.0
numpy==1.24.3
opencv_contrib_python==4.8.1.78
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
pandas==2.1.3
Pillow==10.1.0
Pillow==10.1.0
scikit_learn==1.3.2
scikit-image
torch==2.1.1
torchvision==0.16.1
ultralytics==8.0.222
webcolors==1.13
